+++ geotrek/tourism/tests/test_parsers.py
@@ -305,7 +305,7 @@ class ParserTests(TranslationResetMixin, TestCase):
             with open(filename, 'r') as f:
                 return json.load(f)
 
-        def side_effect(url, allow_redirects, params):
+        def side_effect(url, allow_redirects, headers, params):
             response = requests.Response()
             response.status_code = 503
             response.url = url