+++ docs/changelog.rst
@@ -13,6 +13,7 @@ CHANGELOG
 
 - Add rules fixture on sensitive area (#3470)
 - Change condition on signage & blade to select many of them (#3847)
+- Allow to set headers in requests from Parsers (#3861)
 
 **Documentation**